requirements.txt
Flask == 0.10.1

Flask-Assets
PyExecJS
